requirements.txt
python-dotenv>=1.0.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
TA-Lib>=0.4.28
finnhub-python>=2.4.18
transformers>=4.40.0
torch>=2.1.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyarrow>=14.0.0
alpaca-py>=0.21.0
